Unred the serialized suite: split ProcessManagementViewModel, retry flaky frame capture
Two failures surfaced once the suite ran to completion again (both pre-date this
branch, from #786-era main):
- ProcessManagementViewModel.cs at 570 lines broke the 500-line architecture
  guard — the bundled-fork/crossing-checkbox block (#786) now lives in a partial
  (ProcessManagementViewModel.BundledFork.cs, main file back to 482); issue #789
  tracks any further split
- Issue574RoutingWalkthroughTests failed intermittently on headless CI with
  CaptureRenderedFrame returning null — retry up to 3 times, the same pattern
  Issue776MixedBackendScreenshotTests already uses
