+++ CAP.Avalonia/ViewModels/ProcessManagementViewModel.cs
@@ -55,22 +55,6 @@ public partial class ProcessManagementViewModel : ObservableObject
     private string? _passThroughVersion;
     private double? _passThroughCoreThicknessNm;
 
-    /// <summary>The store forks of bundled PDKs are written to (managed user root).</summary>
-    private readonly UserPdkStore _userPdkStore;
-
-    /// <summary>The fork created by an implicit bundled-source save; later saves write to it directly.</summary>
-    private PdkDraft? _forkedDraft;
-    private string? _forkPath;
-
-    /// <summary>
-    /// Checked when metal traces may cross photonic waveguides directly; unchecked means every
-    /// metal/waveguide crossing needs a bridge. Maps to <see cref="ProcessDefinition.ElectricalBridgeRequired"/>
-    /// inverted — null/absent and false both mean direct crossing, so checked writes null to keep
-    /// unedited files byte-identical (the saver drops nulls).
-    /// </summary>
-    [ObservableProperty]
-    private bool _metalMayCrossPhotonic = true;
-
     /// <summary>The cross-section kinds offered in the editor (Optical / Metal).</summary>
     public static IReadOnlyList<XsectionKind> XsectionKinds { get; } =
         new[] { XsectionKind.Optical, XsectionKind.Metal };
@@ -406,13 +390,6 @@ public partial class ProcessManagementViewModel : ObservableObject
     /// </summary>
     public Func<string, Task<bool>>? ConfirmSaveToPdk { get; set; }
 
-    /// <summary>
-    /// Raised when a save on a bundled read-only PDK implicitly created the user's custom
-    /// copy (fork). The UI layer swaps the library entry to the fork — it keeps the bundled
-    /// name, so the active process keeps resolving it by value.
-    /// </summary>
-    public event EventHandler<BundledPdkForkSavedEventArgs>? BundledPdkForkSaved;
-
     [RelayCommand]
     private async Task SaveProcess()
     {
@@ -487,53 +464,6 @@ public partial class ProcessManagementViewModel : ObservableObject
         }
     }
 
-    /// <summary>
-    /// First save on a bundled read-only PDK: writes the edited process (with all of the
-    /// bundled PDK's components) as the user's fork in the managed root — the bundled JSON
-    /// is never touched — and points the editor at the fork for the rest of the session.
-    /// </summary>
-    private void SaveAsFork(PdkDraft bundledDraft)
-    {
-        try
-        {
-            var edited = ToProcess();
-            // Only rows owned by this PDK — a foreign PDK pulled in for reference via
-            // ImportFromPdk must never be written into the fork either.
-            edited.Layers = edited.Layers.Where(l => _ownedRows.Contains(l)).ToList();
-            edited.Xsections = edited.Xsections.Where(x => _ownedRows.Contains(x)).ToList();
-            edited.Materials = edited.Materials.Where(m => _ownedRows.Contains(m)).ToList();
-            var forkDraft = new PdkDraft
-            {
-                Name = bundledDraft.Name,
-                Description = bundledDraft.Description,
-                Foundry = bundledDraft.Foundry,
-                Version = bundledDraft.Version,
-                DefaultWavelengthNm = bundledDraft.DefaultWavelengthNm,
-                ProcessAgnostic = bundledDraft.ProcessAgnostic,
-                NazcaModuleName = bundledDraft.NazcaModuleName,
-                Backend = bundledDraft.Backend,
-                GdsFactoryRoutingCrossSection = bundledDraft.GdsFactoryRoutingCrossSection,
-                Components = new List<PdkComponentDraft>(bundledDraft.Components),
-                Process = edited,
-            };
-            var forkPath = _userPdkStore.SaveDraftAsFork(forkDraft, bundledDraft.Name);
-            forkDraft.FilePath = forkPath;
-            _forkedDraft = forkDraft;
-            _forkPath = forkPath;
-            _memberDrafts = new List<PdkDraft> { forkDraft };
-
-            BundledPdkForkSaved?.Invoke(this, new BundledPdkForkSavedEventArgs(bundledDraft.Name, forkPath));
-            StatusText = string.Format(
-                LocalizationService.Instance.Translate("ProcessMgmt.Status.SavedAsFork"), Path.GetFileName(forkPath));
-            ProcessSaved?.Invoke(this, EventArgs.Empty);
-        }
-        catch (Exception ex)
-        {
-            StatusText = string.Format(
-                LocalizationService.Instance.Translate("ProcessMgmt.Status.SaveFailed"), ex.Message);
-        }
-    }
-
     /// <summary>Adds an empty material row for manual entry.</summary>
     [RelayCommand]
     private void AddMaterial()
@@ -549,22 +479,4 @@ public partial class ProcessManagementViewModel : ObservableObject
         foreach (var item in items)
             target.Add(item);
     }
-}
-
-/// <summary>Payload of <see cref="ProcessManagementViewModel.BundledPdkForkSaved"/>: the bundled
-/// PDK's name and the fork file the first save wrote in the managed user root.</summary>
-public sealed class BundledPdkForkSavedEventArgs : EventArgs
-{
-    /// <summary>Creates the payload.</summary>
-    public BundledPdkForkSavedEventArgs(string pdkName, string forkPath)
-    {
-        PdkName = pdkName;
-        ForkPath = forkPath;
-    }
-
-    /// <summary>Name of the bundled PDK the fork shadows (kept identical on purpose).</summary>
-    public string PdkName { get; }
-
-    /// <summary>Full path of the written fork JSON in the managed user root.</summary>
-    public string ForkPath { get; }
 }
